+++ zenoh/src/api/builders/mod.rs
@@ -12,6 +12,7 @@
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
 
+pub(crate) mod info;
 pub(crate) mod publisher;
 pub(crate) mod queryable;
 pub(crate) mod reply;
