+++ zenoh/src/api/info.rs
@@ -13,146 +13,10 @@
 //
 
 //! Tools to access information about the current zenoh [`Session`](crate::Session).
-use std::future::{IntoFuture, Ready};
-
-use zenoh_config::wrappers::ZenohId;
-use zenoh_core::{Resolvable, Wait};
-use zenoh_protocol::core::WhatAmI;
-
-use crate::net::runtime::Runtime;
-
-/// A builder returned by [`SessionInfo::zid()`](SessionInfo::zid) that allows
-/// to access the [`ZenohId`] of the current zenoh [`Session`](crate::Session).
-///
-/// # Examples
-/// ```
-/// # #[tokio::main]
-/// # async fn main() {
-///
-/// let session = zenoh::open(zenoh::Config::default()).await.unwrap();
-/// let zid = session.info().zid().await;
-/// # }
-/// ```
-#[must_use = "Resolvables do nothing unless you resolve them using `.await` or `zenoh::Wait::wait`"]
-pub struct ZenohIdBuilder<'a> {
-    runtime: &'a Runtime,
-}
-
-impl<'a> Resolvable for ZenohIdBuilder<'a> {
-    type To = ZenohId;
-}
-
-impl<'a> Wait for ZenohIdBuilder<'a> {
-    fn wait(self) -> Self::To {
-        self.runtime.zid()
-    }
-}
-
-impl<'a> IntoFuture for ZenohIdBuilder<'a> {
-    type Output = <Self as Resolvable>::To;
-    type IntoFuture = Ready<<Self as Resolvable>::To>;
-
-    fn into_future(self) -> Self::IntoFuture {
-        std::future::ready(self.wait())
-    }
-}
-
-/// A builder returned by [`SessionInfo::routers_zid()`](SessionInfo::routers_zid) that allows
-/// to access the [`ZenohId`] of the zenoh routers this process is currently connected to
-/// or the [`ZenohId`] of the current router if this code is run from a router (plugin).
-///
-/// # Examples
-/// ```
-/// # #[tokio::main]
-/// # async fn main() {
-///
-/// let session = zenoh::open(zenoh::Config::default()).await.unwrap();
-/// let mut routers_zid = session.info().routers_zid().await;
-/// while let Some(router_zid) = routers_zid.next() {}
-/// # }
-/// ```
-#[must_use = "Resolvables do nothing unless you resolve them using `.await` or `zenoh::Wait::wait`"]
-pub struct RoutersZenohIdBuilder<'a> {
-    runtime: &'a Runtime,
-}
-
-impl<'a> Resolvable for RoutersZenohIdBuilder<'a> {
-    type To = Box<dyn Iterator<Item = ZenohId> + Send + Sync>;
-}
-
-impl<'a> Wait for RoutersZenohIdBuilder<'a> {
-    fn wait(self) -> Self::To {
-        Box::new(
-            zenoh_runtime::ZRuntime::Application
-                .block_in_place(self.runtime.manager().get_transports_unicast())
-                .into_iter()
-                .filter_map(|s| {
-                    s.get_whatami()
-                        .ok()
-                        .and_then(|what| (what == WhatAmI::Router).then_some(()))
-                        .and_then(|_| s.get_zid().map(Into::into).ok())
-                }),
-        )
-    }
-}
-
-impl<'a> IntoFuture for RoutersZenohIdBuilder<'a> {
-    type Output = <Self as Resolvable>::To;
-    type IntoFuture = Ready<<Self as Resolvable>::To>;
-
-    fn into_future(self) -> Self::IntoFuture {
-        std::future::ready(self.wait())
-    }
-}
-
-/// A builder returned by [`SessionInfo::peers_zid()`](SessionInfo::peers_zid) that allows
-/// to access the [`ZenohId`] of the zenoh peers this process is currently connected to.
-///
-/// # Examples
-/// ```
-/// # #[tokio::main]
-/// # async fn main() {
-///
-/// let session = zenoh::open(zenoh::Config::default()).await.unwrap();
-/// let zid = session.info().zid().await;
-/// let mut peers_zid = session.info().peers_zid().await;
-/// while let Some(peer_zid) = peers_zid.next() {}
-/// # }
-/// ```
-#[must_use = "Resolvables do nothing unless you resolve them using `.await` or `zenoh::Wait::wait`"]
-pub struct PeersZenohIdBuilder<'a> {
-    runtime: &'a Runtime,
-}
-
-impl<'a> Resolvable for PeersZenohIdBuilder<'a> {
-    type To = Box<dyn Iterator<Item = ZenohId> + Send + Sync>;
-}
-
-impl<'a> Wait for PeersZenohIdBuilder<'a> {
-    fn wait(self) -> <Self as Resolvable>::To {
-        Box::new(
-            zenoh_runtime::ZRuntime::Application
-                .block_in_place(self.runtime.manager().get_transports_unicast())
-                .into_iter()
-                .filter_map(|s| {
-                    s.get_whatami()
-                        .ok()
-                        .and_then(|what| (what == WhatAmI::Peer).then_some(()))
-                        .and_then(|_| s.get_zid().map(Into::into).ok())
-                }),
-        )
-    }
-}
-
-impl<'a> IntoFuture for PeersZenohIdBuilder<'a> {
-    type Output = <Self as Resolvable>::To;
-    type IntoFuture = Ready<<Self as Resolvable>::To>;
-
-    fn into_future(self) -> Self::IntoFuture {
-        std::future::ready(self.wait())
-    }
-}
-
+use crate::{
+    api::builders::info::{PeersZenohIdBuilder, RoutersZenohIdBuilder, ZenohIdBuilder},
+    net::runtime::Runtime,
+};
 /// Struct returned by [`Session::info()`](crate::Session::info) which allows
 /// to access information about the current zenoh [`Session`](crate::Session).
 ///
@@ -183,9 +47,7 @@ impl SessionInfo {
     /// # }
     /// ```
     pub fn zid(&self) -> ZenohIdBuilder<'_> {
-        ZenohIdBuilder {
-            runtime: &self.runtime,
-        }
+        ZenohIdBuilder::new(&self.runtime)
     }
 
     /// Return the [`ZenohId`] of the zenoh routers this process is currently connected to
@@ -202,9 +64,7 @@ impl SessionInfo {
     /// # }
     /// ```
     pub fn routers_zid(&self) -> RoutersZenohIdBuilder<'_> {
-        RoutersZenohIdBuilder {
-            runtime: &self.runtime,
-        }
+        RoutersZenohIdBuilder::new(&self.runtime)
     }
 
     /// Return the [`ZenohId`] of the zenoh peers this process is currently connected to.
@@ -220,8 +80,6 @@ impl SessionInfo {
     /// # }
     /// ```
     pub fn peers_zid(&self) -> PeersZenohIdBuilder<'_> {
-        PeersZenohIdBuilder {
-            runtime: &self.runtime,
-        }
+        PeersZenohIdBuilder::new(&self.runtime)
     }
 }
